Update to delete badge actions
I had some issues reconciling the numbers between the PurchasedBadges table and the PurchaseHistory table. I found that the DeleteBadge action and the SuperAdminDeleteBadge action were not recording the details properly in these tables. The DeleteBadge action would record the item as Deleted in the PB table, but not in the PH table, so it created a mismatch when I tried to reconcile the 2 tables. A similar thing happens with the SuperAdminDeleteBadge action. This is my suggested fix.